refactor(agent-guard): move domain guards to skill ownership; keep only universal git guards bundled
Realises the "each skill owns its deterministic guards" model on top of the
agent-guard discovery mechanism (#494):

- agent-guard now bundles only the two universal git hygiene guards
  (commit-trailer, empty-rebase). The engine, GuardContext API, and guards.d
  discovery are unchanged.
- The PR-triage guards (mention, mark-ready) move to
  skills/pr-management-triage/guards/, and the public-PR security-language guard
  to skills/security-issue-fix/guards/ — each an import-free guard(ctx) file
  discovered at runtime; no settings.json re-wiring to add or remove one.
- Setup collects every skills/*/guards/*.py (alongside the engine's bundled
  guards.d) into the adopter .claude/hooks/guards.d/ and the user-scope
  ~/.claude/scripts/guards.d/ — adopt/upgrade/verify + isolated-setup
  install/update updated; docs in secure-agent-setup.md + the tool README's
  Contributing guards contract.

Tests split: test_guards.py covers the engine + bundled guards + discovery;
test_skill_guards.py exercises the relocated guards end-to-end through the real
skills/*/guards dirs via STEWARD_GUARD_DIRS (50 tests; ruff/mypy clean,
validator + workspace-members green).

Generated-by: Claude Code (Opus 4.8 1M context)
